third_party/WebKit/LayoutTests/media/track/opera/interfaces/TextTrackCue/vertical-expected.txt

 
PASS TextTrackCue.vertical, script-created cue 
FAIL TextTrackCue.vertical, parsed cue assert_equals: expected "rl" but got "lr"
